[RMQ.Consumer] Message key added to the consumer context (string for RabbitMQ)
