+++ src/RMQ.Client.Abstractions/Consuming/ConsumerDelegate.cs
@@ -8,6 +8,6 @@ public delegate Task<ProcessResult> ConsumerDelegate<TNativeProperties>(
     ConsumerContext<TNativeProperties> context,
     CancellationToken cancellationToken);
 
-public delegate Task<ProcessResult> ConsumerDelegate<TMessage, TNativeProperties>(
-    ConsumerContext<TMessage, TNativeProperties> context,
+public delegate Task<ProcessResult> ConsumerDelegate<TKey, TMessage, TNativeProperties>(
+    ConsumerContext<TKey, TMessage, TNativeProperties> context,
     CancellationToken cancellationToken);
